+++ gpu/kinfu/src/marching_cubes.cpp
@@ -61,7 +61,7 @@ pcl::gpu::MarchingCubes::run(const TsdfVolume& tsdf, DeviceArray<PointType>& tri
 {  
   if (triangles_buffer.empty())
     triangles_buffer.create(DEFAULT_TRIANGLES_BUFFER_SIZE);
-  occupied_voxels_buffer_.create(3, triangles_buffer.size() / 3);    
+  occupied_voxels_buffer_.create(3, static_cast<int> (triangles_buffer.size () / 3));    
 
   device::bindTextures(edgeTable_, triTable_, numVertsTable_);
